app/requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
anthropic>=0.30.0
google-genai>=1.0.0
python-dotenv>=1.0.0
crawl4ai>=0.5.0
